requirements.txt
Click==7.0
Flask==1.1.2
Flask_Bootstrap==3.3.7.1
Flask_Login==0.5.0
Flask_SQLAlchemy==2.4.4
python_dateutil==2.8.2
